Fixed the tabindex numbers in get-going.html
After rearranging get-going.html, I neglected to update the tabindex for
the various elements.  I picked something here that seems sane to me.

One ordering is visually counter-intuitive: after the opponent's email,
you jump up to the board size, which is at the start of the same column.
However, I found this better than anything else I played with.

The opponent's colour is skipped as it was before.